bball_data3b/requirements.txt
ipykernel
notebook
ipywidgets
ipyevents

hdbscan
json5
matplotlib
mpld3
numpy
orjson
pandas
plotly
pyarrow
pyftdi
pyserial
pyzmq
scipy
seaborn
tiflash
